pyproject.toml
[tool.poetry]
name = "login"
version = "0.1.0"
description = ""
authors = ["Dianyi Jiang  <dianyi.jiang@lincolnuni.ac.nz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.2"
jinja2 = "^3.1.2"
mysql-connector-python = "^8.1.0"
python-dotenv = "^1.0.0"
gitpython = "^3.1.32"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
